@@ -4,9 +4,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "axios": "^0.26.1",
     "gh-pages": "^2.0.1",
-    "react": "^16.6.1",
-    "react-dom": "^16.6.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-scripts": "2.1.1"
   },
   "scripts": {
