@@ -1,23 +1,32 @@
-import React, { Component } from 'react';
+import React, { useState } from 'react';
+import NewTodo from './NewTodo';
 import './Todo.css';
 
-class Todo extends Component {
-  render() {
+const Todo = ({item, deleteTodo, completeTodo}) => {
+
+    const [finished, setFinished] = useState(item.completed)
+
+
     return (
         <div className="todo-container">
             <div
             className='todo'
             >
-                <p>Test Todo Item</p>
+                <p>{item.text}</p>
                 <div
                 className='todo-btn'
                 >
-                <button>Done</button>
+                <button type="button" className={`${finished && 'done'}`}
+                onClick={()=>{
+                    completeTodo(item.id, !item.completed);
+                setFinished(!item.completed)
+                }}>{finished ? 'Done': 'Complete'}</button>
+                <button type="button" onClick={()=>deleteTodo(item.id)}>Delete</button>
                 </div>
             </div>
         </div>   
     );
-  }
+  
 }
 
 export default Todo;
